Refactor cloning test
